+++ packages/calculation/src/index.ts
@@ -1,4 +1,4 @@
 export { Calculation } from "./Calculation";
 export type { CalculationProps } from "./Calculation";
-export { Given, Sum, Difference, Product, Quotient, Ref } from "./elements";
-export type { QuantityProps, GivenProps, OperatorProps, RefProps } from "./elements";
+export { Given, Sum, Difference, Product, Quotient, Ref, Chain, Plus, Minus, Times, DividedBy, Interim } from "./elements";
+export type { QuantityProps, GivenProps, OperatorProps, RefProps, ChainProps, ChainOperandProps } from "./elements";